requirements.txt
asgiref==3.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.0.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.4.0
